% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_coef_path}
\alias{plot_coef_path}
\title{Plot coefficient paths along lambda}
\usage{
plot_coef_path(df, title)
}
\arguments{
\item{df}{Data frame from \code{make_coef_df()}}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
Plot coefficient paths along lambda
}
