% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_coef_paths_compare}
\alias{plot_coef_paths_compare}
\title{Compare coefficient paths across CV methods}
\usage{
plot_coef_paths_compare(res_list)
}
\arguments{
\item{res_list}{Named list of CV results, each with \code{beta_path} and \code{lambda}}
}
\value{
A grob object
}
\description{
Compare coefficient paths across CV methods
}
