% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv_compare.R
\name{plot_cv_compare}
\alias{plot_cv_compare}
\title{Plot CV curves and runtime comparison across methods}
\usage{
plot_cv_compare(res_list)
}
\arguments{
\item{res_list}{Named list of CV results. Each element should contain \code{lambda}, \code{cvm}, and \code{time}.}
}
\value{
A list with \code{plot}, \code{time_table}, and \code{combined_grob}.
}
\description{
Plot CV curves and runtime comparison across methods
}
